Move setupExpressCareMocks to helpers.js
