+++ src/applications/vaos/tests/mocks/setup.js
@@ -6,7 +6,6 @@ import thunk from 'redux-thunk';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import { fireEvent, waitFor } from '@testing-library/dom';
-import moment from 'moment';
 
 import { commonReducer } from 'platform/startup/store';
 import { renderInReduxProvider } from 'platform/testing/unit/react-testing-library-helpers';
@@ -23,21 +22,10 @@ import ExpressCareReasonPage from '../../containers/ExpressCareReasonPage';
 import { cleanup } from '@testing-library/react';
 import ClinicChoicePage from '../../containers/ClinicChoicePage';
 import PreferredDatePage from '../../containers/PreferredDatePage';
-import {
-  getParentSiteMock,
-  getFacilityMock,
-  getExpressCareRequestCriteriaMock,
-} from './v0';
-import {
-  mockParentSites,
-  mockSupportedFacilities,
-  mockRequestEligibilityCriteria,
-  mockRequestLimit,
-} from './helpers';
+import { getParentSiteMock, getFacilityMock } from './v0';
+import { mockParentSites, mockSupportedFacilities } from './helpers';
 
 import createRoutesWithStore from '../../routes';
-
-const today = moment();
 
 export function createTestStore(initialState) {
   return createStore(
@@ -181,44 +169,6 @@ export async function setPreferredDate(store, preferredDate) {
   await cleanup();
 
   return router.push.firstCall.args[0];
-}
-
-export function setupExpressCareMocks({
-  facilityId = '983',
-  isWindowOpen = true,
-  isUnderRequestLimit = true,
-  startTime = null,
-  endTime = null,
-} = {}) {
-  const start =
-    startTime ||
-    today
-      .clone()
-      .subtract(5, 'minutes')
-      .tz('America/Denver');
-  const end =
-    endTime ||
-    today
-      .clone()
-      .add(isWindowOpen ? 3 : -3, 'minutes')
-      .tz('America/Denver');
-  const requestCriteria = getExpressCareRequestCriteriaMock(facilityId, [
-    {
-      day: today
-        .clone()
-        .tz('America/Denver')
-        .format('dddd')
-        .toUpperCase(),
-      canSchedule: true,
-      startTime: start.format('HH:mm'),
-      endTime: end.format('HH:mm'),
-    },
-  ]);
-  mockRequestEligibilityCriteria([facilityId], requestCriteria);
-  mockRequestLimit({
-    facilityId,
-    numberOfRequests: isUnderRequestLimit ? 0 : 1,
-  });
 }
 
 export async function setExpressCareFacility({ store, router }) {
